requirements.txt
mcp[cli]>=1.3.0
requests>=2.32.0
python-dotenv>=1.0.0 
PyYAML>=6.0.1
prance>=23.6.21.0
openapi-spec-validator>=0.7.1
anthropic>=0.20.0
openai>=1.65.3
ruff>=0.2.2
pytest
pytest-xdist
pytest-asyncio
fastapi